01: use pending variable for signals and Arc<Semaphore> read receipts
This allows for new vehicles at the intersection to cross immediately
(with the vehicles themselves doing the sleep(3) in the critical
section), while also avoiding busy waits using a basic semaphore read
receipt setup (with the semaphore shared to allow for waiting for a
single crossing in either lane *and* atomic reference counting to avoid
use-after-frees or complicated locking setups).

All-in-all, it's a bit more complicated but it's much more robust as a
result.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>